ttft_sim/requirements.txt
simpy==4.1.1
numpy==2.0.2
PyYAML==6.0.2
click==8.1.7
matplotlib==3.9.0
